+++ wojilu/ORM/Caching/ContextPool.cs
@@ -93,7 +93,7 @@ namespace wojilu.ORM.Caching {
 
 
 
-        public IEntity FindOne( Type t, int id ) {
+        public IEntity FindOne( Type t, long id ) {
             String key = CacheKey.getObject( t, id );
             return getFromContext( key ) as IEntity;
         }
@@ -149,7 +149,7 @@ namespace wojilu.ORM.Caching {
             return getFromContext( queryKey ) as IList;
         }
 
-        public void Delete( Type t, int id ) {
+        public void Delete( Type t, long id ) {
 
             ContextCache.Remove( t.FullName, id );
         }
