+++ wojilu/ORM/Operation/DeleteOperation.cs
@@ -33,7 +33,7 @@ namespace wojilu.ORM.Operation {
             return Delete( obj.Id, obj, Entity.GetInfo( obj ) );
         }
 
-        public static int Delete( Type t, int id ) {
+        public static int Delete( Type t, long id ) {
             IEntity obj = FindByIdOperation.FindById( id, new ObjectInfo( t ) );
             if (obj != null) {
                 return Delete( obj );
@@ -42,7 +42,7 @@ namespace wojilu.ORM.Operation {
         }
 
 
-        public static int Delete( int id, IEntity obj, EntityInfo entityInfo ) {
+        public static int Delete( long id, IEntity obj, EntityInfo entityInfo ) {
 
             List<IInterceptor> ilist = MappingClass.Instance.InterceptorList;
             for (int i = 0; i < ilist.Count; i++) {
@@ -103,13 +103,13 @@ namespace wojilu.ORM.Operation {
             return DeleteBatch( deleteCondition, Entity.GetInfo( t.FullName ) );
         }
 
-        private static int deleteSingle( int id, EntityInfo entityInfo ) {
+        private static int deleteSingle( long id, EntityInfo entityInfo ) {
 
             IDbCommand command = DataFactory.GetCommand( getDeleteSql( id, entityInfo ), DbContext.getConnection( entityInfo ) );
             return command.ExecuteNonQuery();
         }
 
-        private static String getDeleteSql( int id, EntityInfo entityInfo ) {
+        private static String getDeleteSql( long id, EntityInfo entityInfo ) {
             String str = String.Format( "delete from {0} where Id={1}", entityInfo.TableName, id );
             logger.Info( "Delete(int id): " + str );
             return str;
