+++ wojilu/_wojilu/Tree.cs
@@ -30,7 +30,7 @@ namespace wojilu {
         /// <summary>
         /// 节点的 Id
         /// </summary>
-        int Id { get; set; }
+        long Id { get; set; }
 
         /// <summary>
         /// 节点的名称
@@ -220,7 +220,7 @@ namespace wojilu {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public Node<T> FindParent( int id ) {
+        public Node<T> FindParent( long id ) {
             Node<T> proxy = getById( id );
             if (proxy == null) return default( Node<T> );
             return proxy.getParent();
@@ -231,7 +231,7 @@ namespace wojilu {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public List<Node<T>> FindPath( int id ) {
+        public List<Node<T>> FindPath( long id ) {
 
             List<Node<T>> nodePath = new List<Node<T>>();
 
@@ -267,7 +267,7 @@ namespace wojilu {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public List<Node<T>> FindChildren( int id ) {
+        public List<Node<T>> FindChildren( long id ) {
             Node<T> proxy = getById( id );
             if (proxy == null) return new List<Node<T>>();
             return proxy.getChildren();
@@ -311,7 +311,7 @@ namespace wojilu {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public T GetById( int id ) {
+        public T GetById( long id ) {
             Node<T> proxy = getById( id );
             if (proxy != null) return proxy.getNode();
             return default( T );
@@ -322,7 +322,7 @@ namespace wojilu {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public int GetDepth( int id ) {
+        public int GetDepth( long id ) {
             Node<T> proxy = getById( id );
             if (proxy != null) return proxy.getDepth();
             return 0;
@@ -333,7 +333,7 @@ namespace wojilu {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public T GetParent( int id ) {
+        public T GetParent( long id ) {
             Node<T> proxy = getById( id );
             if (proxy == null) return default( T );
             Node<T> parentProxy = proxy.getParent();
@@ -346,7 +346,7 @@ namespace wojilu {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public List<T> GetPath( int id ) {
+        public List<T> GetPath( long id ) {
 
             List<T> nodePath = new List<T>();
 
@@ -374,7 +374,7 @@ namespace wojilu {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public List<T> GetChildren( int id ) {
+        public List<T> GetChildren( long id ) {
             Node<T> proxy = getById( id );
             if (proxy == null) return new List<T>();
             List<Node<T>> children = proxy.getChildren();
@@ -439,13 +439,13 @@ namespace wojilu {
         /// <param name="nodeId">当select用于设置父节点之时，此参数表示将节点自己排除在下拉列表之外，防止将自己作为自己的父节点</param>
         /// <param name="rootSelectName">根节点(并不存在，但你可以给它取个名称)</param>
         /// <returns></returns>
-        public String DropList( String dropName, int selectValue, int nodeId, String rootSelectName ) {
+        public String DropList( String dropName, long selectValue, long nodeId, String rootSelectName ) {
 
             List<Node<T>> list = this.FindAllOrdered();
 
             StringBuilder builder = new StringBuilder();
             builder.AppendFormat( "<select name=\"{0}\" id=\"{0}\">", dropName );
-            int selval = getSelectedValue( dropName, selectValue );
+            long selval = getSelectedValue( dropName, selectValue );
 
             if (strUtil.HasText( rootSelectName )) {
                 String strSel = selectValue == 0 ? "selected" : "";
@@ -478,7 +478,7 @@ namespace wojilu {
             return builder.ToString();
         }
 
-        private static int getSelectedValue( String dropName, int val ) {
+        private static long getSelectedValue( String dropName, long val ) {
             if (CurrentRequest.getHttpMethod().Equals( "POST" )) {
                 return cvt.ToInt( CurrentRequest.getForm( dropName ) );
             }
@@ -599,7 +599,7 @@ namespace wojilu {
             return builder.ToString();
         }
 
-        private String getListClass( Boolean showChildren, int currentNodeId, Node<T> node ) {
+        private String getListClass( Boolean showChildren, long currentNodeId, Node<T> node ) {
 
             String cls = "";
 
@@ -614,7 +614,7 @@ namespace wojilu {
             return cls.Trim();
         }
 
-        private String getDisplayCls( Boolean showChildren, int currentNodeId, int nodeId ) {
+        private String getDisplayCls( Boolean showChildren, long currentNodeId, long nodeId ) {
 
             if (showChildren) return "";
 
@@ -629,7 +629,7 @@ namespace wojilu {
             return " class=\"hide\"";
         }
 
-        private Boolean isOpenedNode( int currentNodeId, int nodeId, Boolean showChildren ) {
+        private Boolean isOpenedNode( long currentNodeId, long nodeId, Boolean showChildren ) {
 
             if (showChildren) return true;
 
@@ -642,7 +642,7 @@ namespace wojilu {
 
         //----------------------------------------------------------------
 
-        private Node<T> getById( int id ) {
+        private Node<T> getById( long id ) {
 
             if (getIdCache().ContainsKey( id ) == false) return default( Node<T> );
 
@@ -652,14 +652,14 @@ namespace wojilu {
         //----------------------------------------------------------------
 
         private List<Node<T>> _proxyList;
-        private Dictionary<int, Node<T>> _idcache;
+        private Dictionary<long, Node<T>> _idcache;
         private List<Node<T>> _roots;
 
         private List<Node<T>> getNodeList() {
             return _proxyList;
         }
 
-        private Dictionary<int, Node<T>> getIdCache() {
+        private Dictionary<long, Node<T>> getIdCache() {
             return _idcache;
         }
 
@@ -670,7 +670,7 @@ namespace wojilu {
         private void initProxyList() {
 
             _proxyList = new List<Node<T>>();
-            _idcache = new Dictionary<int, Node<T>>();
+            _idcache = new Dictionary<long, Node<T>>();
             _roots = new List<Node<T>>();
 
             // 缓存id
